packages/ai-engine/pyproject.toml
[project]
name = "ai-engine"
version = "0.1.0"
description = "Core AI logic for Awaqi"
requires-python = ">=3.11"
dependencies = [
    "google-genai",
    "pypdf",
    "numpy",
    "database",
    "utils"
]

[tool.uv.sources]
database = { workspace = true }
utils = { workspace = true }
